Move ParticipatePage directory into ParticipateView
