+++ frontend/src/components/MeetingView/ParticipateView/ParticipatePages/ParticipatePage.scss
@@ -1,5 +1,5 @@
-@import '../../utils/colors';
-@import '../../utils/_fonts.scss';
+@import '../../../../utils/colors';
+@import '../../../../utils/_fonts.scss';
 
 .ParticipatePage {
   box-sizing: border-box;
